Issue 17: create new output files each time the pipeline is run
Previously output was appended to existing files. Now, when an
experiment is run, any pre-exiting binary relation output files
are overwitten. Specifically:
* binary_relation.py: amended the process method to ensure that
new empty output files are created for the json and human
readable output. These files are then appended to following the
processing of each batch file (as before).
* helper_functions.py: added create_files function to create
one or more blank files with a specified encoding from a list
of filenames.