Issue 14: Add option to run only a section of the pipeline
Added the "partial_execution" variable to the config file.
If set to False, the complete pipeline will run. If set to
True, a section of the pipeline, corresponding to the
"start_step" and "end_step" values in the config file will
be run.
Note: partial execution is to be used only to re-run steps
of the pipeline (as each step relies on the output of a
previous step).

* Added get_pipeline_steps function to main.py to determine
which steps of the pipeline to run. Added logic to the main
method to ensure that only the requested pipeline steps are
run.
* Added read_group_batches function to helper_functions.py
to read the file specified by the "batch_groups_file"
variable in the config file. This is necessary if the first
step of the pipeline (batching) is not included in the
steps.
* Amended the README.txt file to include notes on the
pipeline steps and corresponding numbers.